- add Fitbit login (and logout) & api integration (no reminders) - update solution server (and app) to use conditions -fix bug in social features where submissions were being duplicated -fix bug where reminders were not showing up -save profile question responses in profile -fix bug where solver crashed on xml comments -update db to remove old tables -remove QuestionResponseOutline class (no longer used) - generic callback interface -fix type in diet.xml - add higher quality icons to main screen - fix time formatter - fix build error when generating signed apk - move activity titles to strings.xml -delete unused node_modules folder

+++ SmartCoach/app/src/main/java/edu/wpi/smartcoach/reminders/ReminderReciever.java
@@ -22,7 +22,7 @@ public class ReminderReciever extends BroadcastReceiver {
 
 	@Override
 	public void onReceive(Context context, Intent intent) {
-		long id = intent.getLongExtra("id", -1);
+		int id = intent.getIntExtra("id", -1);
 		
 		if(id == 9001){//the checkin reminder
 			  Intent checkinIntent = new Intent(context, CheckinActivity.class);
@@ -37,8 +37,7 @@ public class ReminderReciever extends BroadcastReceiver {
 			NotificationManager notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
 
 			notificationManager.notify(0, n);
-		}
-		else if(id != -1) { // a regular reminder
+		} else if(id != -1) { // a regular reminder
 			Reminder r = ReminderService.getInstance().getReminder(id, context);
 			if(r != null){
 				NotificationCompat.Builder builder = new NotificationCompat.Builder(context);
